tests/requirements.txt
# Testing dependencies
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.14.0
pytest-timeout==2.1.0
jsonschema==4.23.0
pyspark==4.0.1
kafka-python==2.0.2
boto3==1.40.67
pyarrow==22.0.0